extract new component CardFilter from room collection filter
