+++ components/Room.tsx
@@ -5,68 +5,22 @@ import { useEffect, useRef, useState } from "react";
 import useMeasure from "react-use-measure";
 import { CalendarRoom } from "./CalendarRoom";
 import { CardCollection } from "./CardCollection";
+import { useFilteredCards, Filters, FilterByReactions } from "./CardFilter";
 import { Discussion } from "./CardView/Discussion";
 import { Desktop } from "./Desktop";
 import { GoToTop, MoreOptionsTiny } from "./Icons";
 import { SearchRoom } from "./SearchRoom";
 import { EditRoomModal } from "./SpaceLayout/Sidebar/RoomListLayout";
 import { RenderedText } from "./Textarea/RenderedText";
-import * as z from "zod";
-import { useSubscribe } from "hooks/useSubscribe";
-import { sortByPosition } from "src/position_helpers";
-
-let FilterVerifier = z.array(
-  z.object({
-    reaction: z.string(),
-    not: z.boolean(),
-  })
-);
-type Filters = z.TypeOf<typeof FilterVerifier>;
 
 export const Room = (props: { entityID: string }) => {
   let roomType = useIndex.eav(props.entityID, "room/type");
   let { ref } = usePreserveScroll<HTMLDivElement>(props.entityID);
 
-  // CARD COLLECTION FILTERS
-  let [filters, setFilters] = useState<Filters>([]);
-
-  // set filter values based on local storage
-  useEffect(() => {
-    try {
-      let filterString = window.localStorage.getItem(
-        `cardCollectionFilters-${props.entityID}`
-      );
-      if (filterString) {
-        let parsed = FilterVerifier.safeParse(JSON.parse(filterString));
-        if (parsed.success) setFilters(parsed.data);
-      }
-    } catch (e) {}
-  }, []);
-  // save filter values to local storage every time the filters state is updated
-  useEffect(() => {
-    window.localStorage.setItem(
-      `cardCollectionFilters-${props.entityID}`,
-      JSON.stringify(filters)
-    );
-  }, [filters]);
-
-  let cards = useCards(props.entityID, "desktop/contains");
-  let reactions = cards.reduce((acc, card) => {
-    for (let reaction of card.reactions) {
-      if (!acc.includes(reaction)) acc.push(reaction);
-    }
-    return acc;
-  }, [] as string[]);
-
-  let cardsFiltered = cards.filter((card) => {
-    let passed = true;
-    for (let filter of filters) {
-      if (filter.not)
-        passed = passed && !card.reactions.includes(filter.reaction);
-      else passed = passed && card.reactions.includes(filter.reaction);
-    }
-    return passed;
-  });
+  let { reactions, filters, setFilters, cardsFiltered } = useFilteredCards(
+    props.entityID,
+    "desktop/contains"
+  );
 
   if (props.entityID === "search") return <SearchRoom />;
   if (props.entityID === "calendar") return <CalendarRoom />;
@@ -276,77 +230,5 @@ function RoomOptions(props: { entityID: string | null }) {
         currentRoom={props.entityID}
       />
     </>
-  );
-}
-
-const useCards = (
-  entityID: string,
-  attribute: "desktop/contains" | "deck/contains"
-) => {
-  let cards = useSubscribe(
-    async (tx) => {
-      let allCards = await scanIndex(tx).eav(entityID, attribute);
-      return Promise.all(
-        allCards.sort(sortByPosition("eav")).map(async (card) => {
-          let reactions = (
-            await scanIndex(tx).eav(card.value.value, "card/reaction")
-          ).map((r) => r.value);
-          return { ...card, reactions };
-        })
-      );
-    },
-    [],
-    [entityID, attribute],
-    `${entityID}-cards`
-  );
-  return cards;
-};
-
-function FilterByReactions(props: {
-  reactions: string[];
-  filters: Filters;
-  setFilters: (f: (old: Filters) => Filters) => void;
-}) {
-  return (
-    <div className="flex flex-row flex-wrap gap-2">
-      {props.reactions.map((reaction) => {
-        let existingFilter = props.filters.find((f) => f.reaction === reaction);
-        return (
-          <button
-            key={reaction}
-            onClick={() => {
-              props.setFilters((oldFilters) => {
-                let existingFilter = oldFilters.find(
-                  (f) => f.reaction === reaction
-                );
-                if (!existingFilter)
-                  return [...oldFilters, { reaction, not: false }];
-                if (existingFilter.not)
-                  return oldFilters.filter((f) => f.reaction !== reaction);
-                return oldFilters.map((f) =>
-                  f.reaction === reaction ? { ...f, not: true } : f
-                );
-              });
-            }}
-            className={`text-md flex items-center gap-2 rounded-md border px-2 py-0.5 ${
-              existingFilter
-                ? existingFilter.not
-                  ? "border-accent-red bg-bg-red"
-                  : "border-accent-green bg-bg-green"
-                : "border-grey-80"
-            }`}
-          >
-            <strong>
-              {existingFilter?.reaction === reaction &&
-                (existingFilter.not ? "−" : "+")}{" "}
-              {reaction}
-            </strong>{" "}
-          </button>
-        );
-      })}
-      {props.filters.length > 0 && (
-        <button onClick={() => props.setFilters(() => [])}>clear</button>
-      )}
-    </div>
   );
 }
